updated whitespace and consolidated componentSize functions
