+++ components/LineupText.js
@@ -1,12 +1,15 @@
 import React, {useState, useCallback} from 'react';
-import {StyleSheet, FlatList, View} from 'react-native';
+import {StyleSheet, FlatList, View, Dimensions} from 'react-native';
 import RecText from '../components/RecText.js'
 import {v4 as uuidv4} from 'uuid'
+
+const width = Dimensions.get('window').width;
+const cardWidth = width-32;
 
 const LineupText = ({lineupData}) => {
     return (
         <View style={styles.flatContain}>
-            <FlatList 
+            <FlatList
                 maxToRenderPerBatch={6}
                 data={lineupData.slice(0,6)}
                 renderItem={({item}) => (
@@ -14,7 +17,7 @@ const LineupText = ({lineupData}) => {
                 )}
                 listKey={uuidv4()}
             />
-            <FlatList 
+            <FlatList
                 maxToRenderPerBatch={6}
                 data={lineupData.slice(6,11)}
                 renderItem={({item}) => (
@@ -33,7 +36,7 @@ const styles = StyleSheet.create({
         paddingVertical: 10,
     },
     itemText: {
-        width: 170,
+        width: (cardWidth/2)-20,
         height: 20,
     }
 });
